First attempt to improve the supervision tree
* I got stuck for some reason I can't start the Stash GenServer from the
  Sequence.Server
   ```elixir
   Supervisor.start_child(sup, worker(Sequence.Stash, [initial_number]))
   ```
