+++ lib/sequence/server.ex
@@ -1,79 +1,19 @@
 defmodule Sequence.Server do
   use GenServer
-  @vsn "1"
-
   require Logger
 
-  defmodule State, do: defstruct current_number: 0, stash_pid: nil, delta: 1
-
-
-  def start_link(stash_pid) do
-    GenServer.start_link(__MODULE__, stash_pid, name: __MODULE__)
+  def start_link(sup, initial_number) do
+    Logger.info "start_link Sequence.Supervisor #{inspect sup} #{inspect initial_number}"
+    GenServer.start_link(__MODULE__, [sup, initial_number])
   end
 
-  def next_number do
-    GenServer.call __MODULE__, :next_number
-  end
-
-  def increment_number(delta) do
-    GenServer.cast __MODULE__, {:increment_number, delta}
-  end
-
-  def get_stash do
-    GenServer.call __MODULE__, :get_stash
-  end
-
-  #####
-  # GenServer implementation
-  def init(stash_pid) do
-    current_number = Sequence.Stash.get_value stash_pid
-    {:ok, %State{current_number: current_number, stash_pid: stash_pid}}
-  end
-
-  def handle_call(:next_number, _from, state) do
-    {
-      :reply,
-      state.current_number,
-      %{state | current_number: state.current_number + state.delta }
-    }
-  end
-
-  def handle_call(:get_stash, _from, state) do
-    {:reply, state.stash_pid, state}
-  end
-
-  def handle_call(_request, _from, state) do
-    {:reply, :ok, state}
-  end
-
-  def handle_cast({:increment_number, delta}, state) do
-    {:noreply,
-      %{state | current_number: state.current_number + delta, delta: delta}
-    }
-  end
-
-  def handle_cast(_msg, state) do
-    {:noreply, state}
-  end
-
-  def handle_info(_info, state) do
-    {:noreply, state}
-  end
-
-  def terminate(_reason, {current_number, stash_pid}) do
-    Sequence.Stash.save_value stash_pid, current_number
-  end
-
-  def format_status(_reason, [ _pdict, state ]) do
-    [data: [{'State', "My current state is '#{inspect state}', and I'm happy"}]]
-  end
-
-  def code_change("0", old_state = { current_number, stash_pid}, _extra) do
-    new_state = %State{current_number: current_number, stash_pid: stash_pid, delta: 1}
-
-    Logger.info "Changing code from 0 to 1"
-    Logger.info inspect(old_state)
-    Logger.info inspect(new_state)
-    {:ok, new_state}
+  def init([sup, initial_number]) do
+    Logger.info "init Sequence.Server #{inspect sup} #{initial_number}"
+    # {:ok, stash} = Supervisor.start_child(sup, worker(Sequence.Stash, [initial_number]))
+    {:ok, stash} = Sequence.Supervisor.start_stash(sup, initial_number)
+    Logger.info "#stash pid {inspect stash}"
+    {:ok, worker_sup} = Sequence.Supervisor.start_worker_sup(sup, stash)
+    Logger.info "#stash pid {inspect worker_sup}"
+    # Supervisor.start_child(sup, supervisor(Sequence.WorkerSupervisor, [stash]))
   end
 end
