+++ lib/sequence/worker_supervisor.ex
@@ -1,4 +1,4 @@
-defmodule Sequence.SubSupervisor do
+defmodule Sequence.WorkerSupervisor do
   use Supervisor
 
   def start_link(stash_pid) do
@@ -6,7 +6,7 @@ defmodule Sequence.SubSupervisor do
   end
 
   def init(stash_pid) do
-    child_processes = [worker(Sequence.Server, [stash_pid])]
+    child_processes = [worker(Sequence.Worker, [stash_pid])]
     supervise child_processes, strategy: :one_for_one
   end
 end
